Grabber & Lift now detecting distance for some actions
Lift is checking bottom distance sensor to see if it has something close to the lift.  Will not lower if something is behind the grabber.

Grabber now has States based on "seeing" things with the distance sensor within a range from the lift bottom.
Grabber will RAISE lift to 2nd junction level setting after it Grabs.
Currently NOT LOWERing the lift automatically when a cone is RELEASED since cones have been accidently bouncing under the grabber near the lift.

In DrivetrainMecanum the handles for the hardware configuration are now explicitly set in the InitDriveTrain().