Python and RC OBJ OpMode for using thresholds
The PipelineStageSwitchingPOWERPLAY.java is an OnBotJava OpMode that uses a webcam to see objects then lets the user cycle through the viewport on the RC phone to see the different types of images in the stages of finding contours using thresholds.  The thresholds still need to be tweaked for the application.  [See python test code].

The python code was used to test the threshold and contour capabilities with OpenCV.  It was able to detect the yellow of junctions, red of a cone and blue of a cone distinctly.  In the conversion the java on the RC phone with EasyOpenCV the thresholds or functions behaved differently.